Allow user-defined exception handling (#24)
To integrate tools like Scout APM, AppSignal, Sentry etc. we need to be able to capture exceptions before converting them to a Twirp error. We add an extra configuration parameter which accepts a Proc, this proc will receive any exception which did not get rescued by the controller rescue_from and did not get re-raised "up the stack". We will assume that if an exception gets re-raised it should be handled by the error handling middleware that APM solutions tend to implement.