Make sure Rack body gets closed in tests (#42)
We have added a middleware to our main app which uses a Rack body proxy. It turns out that the test methods from this gem would not call `close` on the Rack body, which leads to resource leaks from the Rails stack (we would be setting a thread-local variable in the Rack action which gets disposed of in the body close).

While it would be nicer to reuse the Rails integration test methods here for the moment let's patch this by making sure we call `close` ourselves.